backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.3.0",
    "google-genai>=1.30.0",
    "langextract>=1.0.8",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.0",
    "yt-dlp>=2025.8.11",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
